requirements.txt
transformers>=5.3.0
torch>=2.0.0
torchaudio>=2.0.0
safetensors>=0.4.0
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0
einops>=0.8.0
accelerate>=1.0.0
peft>=0.13.0
openai-whisper
tensorboard>=2.14.0
